requirements.txt
dnspython
tgcrypto
python-dotenv
flask
aiofiles
aiohttp
uvloop
psutil
tgcrypto
aiofiles
aiohttp
dnspython
flask
psutil
python-dotenv
uvloop
pyaes
pysocks
aiofiles
aiohttp
dnspython
flask
psutil
pyaes
pyrogram
pysocks
python-dotenv
tgcrypto
uvloop
